add: :art: #2 categorytitlewidgetを別ファイル化 ~11:51
